pyproject.toml
[tool.poetry]
name = "test-task-salesbrush"
version = "0.1.0"
description = ""
authors = ["sauberr <sauberr10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.12.0"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"
typer = "^0.20.0"
apscheduler = "^3.11.1"
loguru = "^0.7.3"
pytest = "^9.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.7"
mypy = "^1.19.0"
pre-commit = "^4.5.0"
pytest = "^9.0.1"
pytest-cov = "^7.0.0"


[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.pyc",
    ".mypy_cache",
    ".pytest_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "SIM",
    "RUF",
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"


[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true
strict_equality = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_any_generics = true
no_implicit_optional = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false
disallow_untyped_calls = false
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "loguru",
    "pydantic_settings",
    "sqlalchemy.*",
    "apscheduler.*",
    "typer",
    "rich.*",
    "pytest",
]
ignore_missing_imports = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
